sysfs: use uint16_t for core count in the insanely high chance a device has more than 255 cores

+++ sysfs.cpp
@@ -13,7 +13,7 @@
 #include "misc.h"
 #include "sysfs.h"
 
-std::vector<uint32_t> SysFs::Cpu::get_freqs(uint8_t core) {
+std::vector<uint32_t> SysFs::Cpu::get_freqs(uint16_t core) {
 	// read and split avail freqs
 	std::string freq_list_str = IO::read_file(PATH_CPU + "/cpu" + std::to_string(core) + "/cpufreq/scaling_available_frequencies");
 	std::istringstream iss(freq_list_str);
